refactor: 拆分预警配置与观察清单模块
