+++ tests/contract_checks.ps1
@@ -164,8 +164,10 @@ Assert-Contains -Path "goldmonitor\alert_rules.py" -Pattern 'def batch_update_al
 Assert-Contains -Path "goldmonitor\socket_alert_rules.py" -Pattern '@socketio\.on\("batch_update_alert_rules"\)' -Message "backend must expose alert rule batch updates"
 Assert-Contains -Path "goldmonitor\socket_alert_rules.py" -Pattern '@socketio\.on\("get_alert_rule_insight"\)' -Message "backend must expose per-rule insight"
 Assert-Contains -Path "goldmonitor\socket_alert_rules.py" -Pattern '@socketio\.on\("simulate_alert_rule"\)' -Message "backend must expose alert rule history simulation"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("save_alert_profile"\)' -Message "backend must expose alert profile save event"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("apply_alert_profile"\)' -Message "backend must expose alert profile apply event"
+Assert-Contains -Path "goldmonitor\socket_alert_configuration.py" -Pattern '@socketio\.on\("set_threshold"\)' -Message "backend must expose threshold updates"
+Assert-Contains -Path "goldmonitor\socket_alert_configuration.py" -Pattern '@socketio\.on\("set_volatility"\)' -Message "backend must expose volatility updates"
+Assert-Contains -Path "goldmonitor\socket_alert_configuration.py" -Pattern '@socketio\.on\("save_alert_profile"\)' -Message "backend must expose alert profile save event"
+Assert-Contains -Path "goldmonitor\socket_alert_configuration.py" -Pattern '@socketio\.on\("apply_alert_profile"\)' -Message "backend must expose alert profile apply event"
 Assert-Contains -Path "goldmonitor\storage_manifest.py" -Pattern 'alert_profiles' -Message "storage manifest must include alert profiles"
 Assert-Contains -Path "goldmonitor\storage_manifest.py" -Pattern 'daily_digest_state' -Message "storage manifest must include daily digest state"
 Assert-Contains -Path "goldmonitor\storage_manifest.py" -Pattern 'review_notes' -Message "storage manifest must include review notes"
@@ -357,26 +359,29 @@ Assert-Contains -Path "templates\index.html" -Pattern 'alertRuleCenterList' -Mes
 Assert-Contains -Path "templates\index.html" -Pattern 'alertRuleSummary' -Message "frontend must render alert rule status summary"
 Assert-Contains -Path "templates\index.html" -Pattern 'alertRuleFilters' -Message "frontend must expose alert rule type filters"
 Assert-Contains -Path "templates\index.html" -Pattern 'alertProfilesPanel' -Message "frontend must render alert profile panel"
+Assert-Contains -Path "templates\index.html" -Pattern 'alert-configuration-center\.js' -Message "frontend must load the alert configuration module"
 Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'renderAlertRuleCenter' -Message "frontend must render unified alert rules"
-Assert-Contains -Path "static\app.js" -Pattern 'alert_profiles_updated' -Message "frontend must handle alert profile updates"
+Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'registerAlertRuleSocketHandlers' -Message "frontend must register unified alert rule events"
+Assert-Contains -Path "static\alert-configuration-center.js" -Pattern 'registerAlertConfigurationSocketHandlers' -Message "frontend must register alert configuration events"
+Assert-Contains -Path "static\alert-configuration-center.js" -Pattern 'alert_profiles_updated' -Message "frontend must handle alert profile updates"
 Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'saveUnifiedAlertRule' -Message "frontend must save unified alert rules"
 Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'duplicateUnifiedAlertRule' -Message "frontend must duplicate unified alert rules"
-Assert-Contains -Path "static\app.js" -Pattern 'alertRuleDraft = rule \? cloneAlertRuleDraft\(rule\)' -Message "duplicated rules must open with a populated edit draft"
+Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'alertRuleDraft = rule \? cloneAlertRuleDraft\(rule\)' -Message "duplicated rules must open with a populated edit draft"
 Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'changeAlertRuleCooldownMode' -Message "frontend must switch per-rule cooldown modes without losing the draft"
 Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'changeAlertRuleDeliveryMode' -Message "frontend must switch per-rule delivery modes without losing the draft"
 Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'alertRuleChannel_' -Message "alert rules must support per-rule notification channels"
-Assert-Contains -Path "static\app.js" -Pattern 'alert_rule_toggled' -Message "frontend must confirm alert rule toggle completion"
-Assert-Contains -Path "static\app.js" -Pattern 'alert_rule_reset' -Message "frontend must confirm alert rule reset completion"
+Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'alert_rule_toggled' -Message "frontend must confirm alert rule toggle completion"
+Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'alert_rule_reset' -Message "frontend must confirm alert rule reset completion"
 Assert-Contains -Path "goldmonitor\socket_alert_rules.py" -Pattern '@socketio\.on\("save_alert_rule"\)' -Message "backend must expose unified alert rule save"
 Assert-Contains -Path "app.py" -Pattern 'check_alert_rules' -Message "price refresh must evaluate unified alert rules"
 Assert-Contains -Path "app.py" -Pattern 'WATCH_TARGETS_PATH' -Message "app.py must persist watch targets locally"
 Assert-Contains -Path "app.py" -Pattern 'load_watch_targets' -Message "app.py must load watch targets from disk"
 Assert-Contains -Path "app.py" -Pattern 'save_watch_targets' -Message "app.py must save watch targets to disk"
 Assert-Contains -Path "app.py" -Pattern 'check_watch_targets' -Message "app.py must check watch targets during price refresh"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("set_watch_target"\)' -Message "backend must create or update watch targets"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("delete_watch_target"\)' -Message "backend must delete watch targets"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("toggle_watch_target"\)' -Message "backend must toggle watch targets"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("reset_watch_target"\)' -Message "backend must reset watch target trigger state"
+Assert-Contains -Path "goldmonitor\socket_alert_configuration.py" -Pattern '@socketio\.on\("set_watch_target"\)' -Message "backend must create or update watch targets"
+Assert-Contains -Path "goldmonitor\socket_alert_configuration.py" -Pattern '@socketio\.on\("delete_watch_target"\)' -Message "backend must delete watch targets"
+Assert-Contains -Path "goldmonitor\socket_alert_configuration.py" -Pattern '@socketio\.on\("toggle_watch_target"\)' -Message "backend must toggle watch targets"
+Assert-Contains -Path "goldmonitor\socket_alert_configuration.py" -Pattern '@socketio\.on\("reset_watch_target"\)' -Message "backend must reset watch target trigger state"
 Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'kind === ''watch_target''' -Message "unified alert center must render watch target rules"
 Assert-Contains -Path "static\alert-log-center.js" -Pattern 'renderNotificationBadges' -Message "alert log must render notification delivery status"
 Assert-Contains -Path "goldmonitor\socket_alert_log.py" -Pattern '@socketio\.on\("export_alert_log"\)' -Message "backend must export alert logs"
